requirements.txt
absl-py==1.0.0
addict==2.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.1.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
easydict==1.11
editdistance==0.6.0
entrypoints==0.4
fastjsonschema==2.15.3
fiona==1.9.5
flatbuffers==23.5.26
fonttools==4.25.0
gast==0.4.0
geopandas==0.10.2
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
grpcio==1.45.0
h5py==3.8.0
hiddenlayer==0.3
idna==3.3
imageio==2.18.0
imgviz==1.5.0
importlib-metadata==6.7.0
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.3.0
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.11.0
kiwisolver @ file:///C:/ci/kiwisolver_1644944417636/work
labelImg==1.8.6
libclang==16.0.6
lmdb==1.4.1
lxml==4.8.0
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.2.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mmengine==0.9.0
munkres==1.1.4
natsort==8.1.0
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
netron==5.7.2
networkx==2.6.3
notebook==6.4.11
numba==0.56.0rc1
numpy @ file:///D:/bld/numpy_1649585408897/work
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow @ file:///D:/bld/pillow_1648857310353/work
platformdirs==3.11.0
plotly==5.7.0
Polygon3==3.0.9.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.6
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycparser==2.21
pydotz==1.5.1
Pygments==2.16.1
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyproj==3.2.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
QtPy==2.1.0
rapidfuzz==2.1.3
regex==2023.10.3
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
Shapely==1.8.2
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
tabulate==0.8.10
tenacity==8.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorwatch==0.9.1
termcolor==1.1.0
terminado==0.13.3
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tomli==2.0.1
torch==1.10.0
torchaudio==0.10.0
torchsummary==1.5.1
torchvision==0.11.0
torchviz==0.0.2
tornado @ file:///C:/ci/tornado_1606935947090/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==5.1.1
typing_extensions==4.7.1
ujson==5.7.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.15.0
yacs==0.1.8
yapf==0.40.2
zipp==3.8.0
